+++ inc/Feature.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Airfleet\Framework;
+
+interface Feature {
+	/**
+	 * Setup.
+	 *
+	 * @return void
+	 */
+	public function initialize(): void;
+}