+++ CHANGELOG.md
@@ -9,5 +9,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
-
 - Created package
+- Added Options utilities